# Conflicts:
#	app/Actions/Blog/StoreBlogAction.php
#	app/Actions/Blog/UpdateBlogAction.php
#	app/Actions/Book/StoreBookAction.php
#	app/Actions/Book/UpdateBookAction.php
#	app/Actions/Faq/UpdateFaqAction.php
#	app/Actions/Translation/SetTranslationAction.php
#	app/Enums/PermissionEnum.php
#	app/Http/Controllers/Api/V1/BlogController.php
#	app/Http/Controllers/Api/V1/BookController.php
#	app/Http/Resources/BlogResource.php
#	app/Http/Resources/CategoryResource.php
#	app/Models/Book.php
#	app/Models/Category.php
#	app/Models/Notice.php
#	app/Models/User.php
#	app/Policies/BlogPolicy.php
#	app/Providers/RepositoryProvider.php
#	app/Providers/RepositoryServiceProvider.php
#	app/Repositories/Category/CategoryRepositoryInterface.php
#	app/Traits/HasSlug.php
#	database/migrations/2023_11_16_142926_create_publications_table.php
#	database/seeders/DatabaseSeeder.php
#	database/seeders/UserSeeder.php
#	lang/en/category.php
#	routes/api/blog.php

+++ app/Actions/Blog/UpdateBlogAction.php
@@ -5,6 +5,7 @@ namespace App\Actions\Blog;
 use App\Actions\Translation\SetTranslationAction;
 use App\Models\Blog;
 use App\Repositories\Blog\BlogRepositoryInterface;
+use App\Repositories\Category\CategoryRepositoryInterface;
 use Illuminate\Support\Facades\DB;
 use Lorisleiva\Actions\Concerns\AsAction;
 
@@ -12,7 +13,11 @@ class UpdateBlogAction
 {
     use AsAction;
 
-    public function __construct(private readonly BlogRepositoryInterface $repository)
+    public function __construct(
+        private readonly BlogRepositoryInterface $repository,
+        private readonly CategoryRepositoryInterface $categoryRepository,
+
+    )
     {
     }
 
@@ -22,18 +27,22 @@ class UpdateBlogAction
      * @param array{name:string,mobile:string,email:string} $payload
      * @return Blog
      */
-    public function handle(Blog $blog, array $payload): Blog
+    public function handle(Blog $blog, array $payload): Blog|null
     {
         return DB::transaction(function () use ($blog, $payload) {
-            $payload['user_id'] = auth()->id();
-            $blog->media()->delete();
-            $blog->update($payload);
-            SetTranslationAction::translate($blog, $payload['translations']);
-            if (request()->hasFile('media')) {
-                $blog->addMediaFromRequest('media')
-                     ->toMediaCollection('blog');
+            $category = $this->categoryRepository->find($payload['category_id']);
+            if ($category->type === Blog::class) {
+                $payload['user_id'] = auth()->id();
+                $blog->media()->delete();
+                $blog->update($payload);
+                SetTranslationAction::translate($blog, $payload['translations']);
+                if (request()->hasFile('media')) {
+                    $blog->addMediaFromRequest('media')
+                         ->toMediaCollection('blog');
+                }
+                return $blog;
             }
-            return $blog;
+            return null;
         });
     }
 }
